requirements.txt
azure-cognitiveservices-speech
openai
python-dotenv
fastapi
uvicorn
python-multipart
sounddevice
soundfile
numpy
pydantic
